Create confirmation for comment deletion (to fix)

+++ blog/views.py
@@ -66,10 +66,17 @@ class PostDetail(View):
 
 
 class CommentDelete(View):
+    def get(self, request, comment_id, *args, **kwargs):
+        comment = get_object_or_404(Comment, id=comment_id)
+        return render(request, 'comment_delete.html', {"comment": comment})
+
     def post(self, request, comment_id, *args, **kwargs):
         comment = get_object_or_404(Comment, id=comment_id)
+
         if request.user.is_superuser or request.user.username == comment.name:
-            comment.delete()
+            if request.POST.get("confirm_delete"):
+                comment.delete()
+            return HttpResponseRedirect(reverse('post_detail', args=[comment.post.slug]))
         else:
             return HttpResponseRedirect(reverse('post_detail', args=[comment.post.slug]))
